Adding information about Chef cookbook
Thanks for your work on Jently. We really appreciate your efforts at Paydici and needed to install it via Chef. I could not find an existing cookbook so we rolled our own. I hope you are cool with adding this information to the README. It only supports ubuntu/debian for now but we are open to pull requests.

Thanks again!